requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
supabase==2.7.4
python-dotenv==1.0.1
requests==2.32.3
